Migration and seed data for test

+++ Crud_API/Controllers/CourseController.cs
@@ -6,6 +6,7 @@ using Crud_API.Interfaces.Repositories;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using System.Linq.Expressions;
 using System.Runtime.InteropServices;
 
@@ -33,7 +34,7 @@ namespace Crud_API.Controllers
 
         }
 
-        [HttpGet, Route("{:id}")]
+        [HttpGet, Route("")]
         public async Task<GenericResult<Course>> GetCourseByIdAsync(long id)
         {
             var result = await _courseRepository.GetAsync(id);
@@ -41,7 +42,10 @@ namespace Crud_API.Controllers
             {
                 return GenericResult<Course>.NotFound;
             }
-            return GenericResult<Course>.Succeed(result);
+            else
+            {
+                return GenericResult<Course>.Succeed(result);
+            }
         }
 
         [HttpPost, Route("create")]
@@ -52,49 +56,87 @@ namespace Crud_API.Controllers
             {
                 return GenericResult<Course>.BadRequest;
             }
+            else
+            {
+                return GenericResult<Course>.CreatedWithMessage("Entity Created Successfully");
+                _logger.LogInformation($"{DateTime.Now}: Course {course.CourseName} added to the database.");
+            }
 
-            return GenericResult<Course>.CreatedWithMessage("Entity Created Successfully");
-            _logger.LogInformation($"{DateTime.Now}: Course {course.CourseName} added to the database.");
         }
 
-        [HttpPut, Route("update/{:id}")]
-        public async Task<GenericResult<Course>> UpdateCourseAsync(long courseId)
+        [HttpPut, Route("update")]
+        public async Task<GenericResult<Course>> UpdateCourseAsync(Course course)
         {
-            var result = await _courseRepository.GetAsync(courseId);
+            var result = await _courseRepository.GetAsync(course.Id);
             if (result == null)
             {
                 return GenericResult<Course>.NotFound;
             }
-            var course = await _courseRepository.Update(result);
-            if (course == null)
+            else
             {
-                return GenericResult<Course>.BadRequest;
+                var updatedCourse = await _courseRepository.Update(course);
+                return GenericResult<Course>.SucceedWithMessage("Entity Successfully Updated");
+                _logger.LogInformation($"{DateTime.Now}: Course {course.CourseName} updated to the database.");
             }
-
-            return GenericResult<Course>.SucceedWithMessage("Entity Successfully Updated");
-            _logger.LogInformation($"{DateTime.Now}: Course {course.CourseName} updated to the database.");
+    
         }
 
-        [HttpDelete, Route("{:id}")]
-        public async Task<IActionResult> Delete(long courseId) 
+        [HttpDelete, Route("delete")]
+        public async Task<GenericResult<bool>> Delete(long courseId)
         {
             var result = await _courseRepository.GetAsync(courseId);
-            return await _courseRepository.Remove(result);
+            if (result == null)
+            {
+                return GenericResult<bool>.NotFound;
+            }
+            else
+            {
+                await _courseRepository.Remove(result);
+                return GenericResult<bool>.SucceedWithMessage("Entity deleted successfully");
+            }
         }
 
-        [HttpGet, Route("filter/{filter}")]
-        public async Task<GenericResult<Course>> GetListAsync(Expression<Func<Course, bool>> filter = null)
+        [HttpGet, Route("filter")]
+        public async Task<GenericResult<IEnumerable<Course>>> GetListAsync([FromQuery] string filter = null)
         {
-            var courses = await _courseRepository.GetListAsync(filter);
-
-            return GenericResult<Course>.Succeed(courses);
+            try
+            {
+                var courses = await _courseRepository.GetListAsync(s => s.CourseName == filter);
+                return GenericResult<IEnumerable<Course>>.Succeed(courses);
+            }
+            catch (Exception ex)
+            {
+                return GenericResult<IEnumerable<Course>>.Fail(ex.Message);
+            }
         }
 
-        [HttpGet, Route("order/{orderby}")]
-        public async Task<GenericResult<Course>> GetSortedAsync(Expression<Func<Course, object>> orderBy, bool ascending)
+        [HttpGet, Route("order")]
+        public async Task<GenericResult<IEnumerable<Course>>> GetSortedAsync([FromQuery] string orderBy, bool ascending)
         {
-            var students = await _courseRepository.GetSortedAsync(orderBy, ascending);
-            return GenericResult<Course>.Succeed(students);
+            try
+            {
+                var courses = await _courseRepository.GetSortedAsync(s => EF.Property<object>(s, orderBy), ascending);
+                return GenericResult<IEnumerable<Course>>.Succeed(courses);
+            }
+            catch (Exception ex)
+            {
+                return GenericResult<IEnumerable<Course>>.Fail(ex.Message);
+            }
+
+        }
+
+        [HttpGet("page")]
+        public async Task<GenericResult<IEnumerable<Course>>> GetPaged([FromQuery] int pageNumber = 1, [FromQuery] int pageSize = 10)
+        {
+            try
+            {
+                var courses = await _courseRepository.GetPagedAsync(pageNumber, pageSize);
+                return GenericResult<IEnumerable<Course>>.Succeed(courses);
+            }
+            catch (Exception ex)
+            {
+                return GenericResult<IEnumerable<Course>>.Fail(ex.Message);
+            }
         }
     }
 }
